Attempting to move the application to Docker containers
