+++ .dockerignore
@@ -1,0 +1,4 @@
+node_modules
+Dockerfile
+.dockerignore
+docker-compose.yml
