+++ .env
@@ -1,11 +1,16 @@
 PORT=5000
-MONGODB_DB_URL=mongodb://127.0.0.1:27017/
-MONGODB_DB_NAME=studentsproject
-MONGODB_USERS_COLLECTION_NAME=users
 JWT_ACCESS_SECRET_KEY=jwt-access-secret-key
-DB_NAME=studentsproject
-DB_USER=postgres
-DB_PASSWORD=postgre
-DB_HOST=localhost
-DB_PORT=5432
-DATABASE_URL=postgres://postgres:postgre@localhost:5432/studentsproject
+MONGODB_DB_NAME=studentsproject
+#MONGODB_DB_HOST=127.0.0.1
+MONGODB_DB_HOST=mongo
+MONGODB_DB_PORT=27017
+MONGODB_USERS_COLLECTION_NAME=users
+REDIS_DB_PORT=6379
+#REDIS_DB_HOST=127.0.0.1
+REDIS_DB_HOST=redis
+POSTGRES_DB_NAME=studentsproject
+POSTGRES_DB_USER=user
+POSTGRES_DB_PASSWORD=password
+#POSTGRES_DB_HOST=127.0.0.1
+POSTGRES_DB_HOST=postgres
+POSTGRES_DB_PORT=5432
