+++ package.json
@@ -1,13 +1,13 @@
 {
   "name": "studentsproject",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Repository for the implementation of the project described in the [technical task](https://docs.google.com/document/d/1smlvEuArov1OVh3S1IOfy0A-nXYQb2F3TohCUPteiMI/edit?usp=sharing).",
   "main": "build/server.js",
   "type": "commonjs",
   "scripts": {
-    "start": "npm run build && node build/app.js",
-    "build": "rimraf ./build && npx tsc",
-    "dev": "npx nodemon",
+    "start": "node ./build/app.js",
+    "dev": "nodemon -L -e ts --exec \"npm run build && npm start\"",
+    "build": "tsc",
     "migrate": "node-pg-migrate",
     "migrate:up": "node-pg-migrate up",
     "migrate:down": "node-pg-migrate down",
@@ -53,7 +53,6 @@
     "joi": "^17.7.0",
     "jsonwebtoken": "^8.5.1",
     "mongodb": "^4.13.0",
-    "mongoose": "^6.8.0",
     "node-pg-migrate": "^6.2.2",
     "pg": "^8.8.0",
     "redis": "^4.6.4",
